chore: move pox2-event parsing code into separate file
